[FEATURE] Move to namespacing
